Polish Combat
Double attack distance and movement distance of all units
For now, the aim button and ram button will only appear if the Orge can be aimed or rammed.
Aim: Click Aim button and choose the component of Orge that you want to aim on the panel.
Ram: Click Ram button then the unit will automatically ram the Orge.